internal: move packages from pkg directory
